==> analysis/03_moderators.Rmd <==
---
title: "03_moderators"
author: "anjie"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
    number_sections: yes
    theme: paper
    toc: yes
    toc_float: yes
---

```{r}
library(tidyverse)
library(here)
library(ggthemes)
library(metafor)
library(kableExtra)
library(ggrepel)

d <- read_csv(here("data/clean_data.csv"))
author_d <- read_csv(here("data/clean_author.csv"))

source(here("helper/model_comparison_help.r"))
source(here("helper/method_mod_help.r"))
source(here("helper/author_help.r"))



best_fit_bmeasure_model_df <- readRDS(here("cached_data/best_fit_behavioral_measure_df.Rds"))
best_fit_exposure_phase_df <- readRDS(here("cached_data/best_fit_exposure_phase.Rds"))
best_fit_real_visual_df <- readRDS(here("cached_data/best_fit_real_visual_df.Rds"))
best_fit_real_aud_df <- readRDS(here("cached_data/best_fit_real_aud_df.Rds"))
best_fit_major_author_df <- readRDS(here("cached_data/best_fit_major_author_df.Rds"))

age_modesl_df <- readRDS(here("cached_data/age_models_df.Rds"))

major_author_fit_df <- readRDS(here("cached_data/major_author_model_ds.Rds"))


```





# 1. behavioral_measure 


## running  models for comparison

```{r eval=FALSE, include=FALSE}

# getting the subset of data
bmeasure_ds <- d %>% 
  filter(!response_mode %in% c("NIRS", "EEG")) %>% 
  distinct(ds_clean, behavioral_measure) %>% 
  group_by(ds_clean) %>% 
  count() %>% 
  filter(!n == 1) %>% 
  pull(ds_clean)

# for model fits


bmeasure_models_df <- lapply(bmeasure_ds, 
       function(name){
         print(name)
         get_compare_IC_df(d, name, c("behavioral_measure"))
       }) %>% 
  bind_rows() %>% 
  mutate(model_type = "bmeasure")


# finding the best fit models when adding method moderator 
min_AICc_bmeasure_df <- bmeasure_models_df %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) 

# finding the stats of the models 
bmeasure_models_fits_df <- lapply(bmeasure_ds, 
       function(name){
         print(name)
         get_model_fit_df(d, name, c("behavioral_measure"))
       }) %>% 
  bind_rows()


# finding the best fit behavioral model 
best_fit_bmeasure_model_df <- bmeasure_models_fits_df %>% 
  left_join(min_AICc_bmeasure_df, 
            by =  c("model_spec", "dataset", "model_spec_clean")) %>% 
  filter(!is.na(REML)) 

saveRDS(best_fit_bmeasure_model_df, here("cached_data/best_fit_behavioral_measure_df.Rds"))
```


## model stats for significant values

```{r}
best_fit_bmeasure_model_df %>% 
  filter(grepl("behavioral_measure", term)) %>% 
  filter(p.value < .05) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```

## visualization 

```{r}
d %>% 
  filter(mean_age_months < 36) %>% 
  filter(ds_clean %in% unique(best_fit_bmeasure_model_df$dataset)) %>% 
  ggplot(aes(x = mean_age_months, y = d_calc, color = behavioral_measure)) + 
  geom_point(alpha = .5) + 
  geom_smooth(method = "lm") + 
  facet_wrap(~ds_clean)+ 
  theme_few()
```



# 2. Exposure phase 

## running models for comparison

```{r eval=FALSE, include=FALSE}
## get subset of data
ep_d <-d %>% 
  filter(!response_mode %in% c("NIRS", "EEG")) %>% 
  filter(exposure_phase != "test_only") %>% 
  distinct(ds_clean, exposure_phase) %>%
  group_by(ds_clean) %>% 
  count() %>% 
  filter(!n == 1) %>% 
  filter(ds_clean != "Natural speech preference") %>% 
  pull(ds_clean)


# taking out the natural speech preference because only 1 ES has certain values 
ep_new_d <- d %>% mutate(exposure_phase = case_when(
  exposure_phase == "familiarization" ~ "a_familiarization", 
  TRUE ~ exposure_phase)) %>% 
  filter(exposure_phase != "test_only") %>% 
  filter(ds_clean != "Natural speech preference")


## get model comparison stats
ep_models_df <- lapply(ep_d, 
       function(name){
         print(name)
         get_compare_IC_df(ep_new_d, name, c("exposure_phase"))
       }) %>% 
  bind_rows() %>% 
  mutate(model_type = "ep")

## find best fitting models 
min_AICc_ep_models_df <- ep_models_df %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) 

## finding the stats of the models 
ep_models_fits_df <- lapply(ep_d, 
       function(name){
         print(name)
         get_model_fit_df(ep_new_d, name, c("exposure_phase"))
       }) %>% 
  bind_rows()

## find the best fit model's stats
best_fit_ep_model_df <- ep_models_fits_df %>% 
  left_join(min_AICc_ep_models_df, 
            by =  c("model_spec", "dataset", "model_spec_clean")) %>% 
  filter(!is.na(REML)) 

saveRDS(best_fit_ep_model_df, here("cached_data/best_fit_exposure_phase.Rds"))
```

## model stats for significant values

```{r}
best_fit_exposure_phase_df %>% 
  filter(grepl("exposure_phase", term)) %>% 
  filter(p.value < .05) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```

## visualization

```{r}
d %>% 
  filter(mean_age_months < 36) %>% 
  filter(ds_clean %in% unique(best_fit_exposure_phase_df$dataset)) %>% 
  ggplot(aes(x = mean_age_months, y = d_calc, color = exposure_phase)) + 
  geom_point(alpha = .5) + 
  geom_smooth(method = "lm") + 
  facet_wrap(~ds_clean)+ 
  theme_few()
```



# 3. Realistic Stimuli 


## visual 

### getting subset 

```{r}
# a slightly hacky way to flip the level to make it look more consistent 

visual_stimulus_d <- bind_rows((d %>% 
  filter(!is.na(object_stimulus)) %>% 
  #select(ds_clean, object_stimulus) %>% 
  mutate(
    visual_stimulus_type = case_when(
      object_stimulus %in% c("drawings", "digital", "paper",
                             "picture", "word") ~ "a_schematic_rep", 
      TRUE ~ "b_real"
    )
  )), (d %>% 
  filter(!is.na(stimuli_type), ds_clean == "Prosocial agents") %>% 
  mutate(visual_stimulus_type = case_when(
    stimuli_type == "cartoon" ~ "a_schematic_rep",
    TRUE ~ "b_real"
  ))))


```



### running models 

```{r eval=FALSE, include=FALSE}


best_fit_visual_stimuli_df <- get_best_fit_model_with_mod(visual_stimulus_d, "visual_stimulus_type")
saveRDS(best_fit_visual_stimuli_df, here("cached_data/best_fit_real_visual_df.Rds"))

```


### significant models  

```{r}
best_fit_real_visual_df %>% 
  filter(grepl("visual_stimulus", term)) %>% 
  filter(p.value < 0.05) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```


###  visualization

```{r}
visual_stimulus_d %>% 
  filter(mean_age_months < 36) %>% 
  ggplot(aes(x = mean_age_months, y = d_calc, color = visual_stimulus_type)) + 
  geom_point(alpha = .5) + 
  geom_smooth(method = "lm") + 
  facet_wrap(~ds_clean)+ 
  theme_few()
```

## auditory 

### getting subset 

```{r}
auditory_stimulus_d <- bind_rows(
  d %>% 
  filter(!is.na(stimuli_type)) %>% 
  filter(ds_clean == "Statistical word segmentation") %>% 
  mutate(auditory_stimulus_type = stimuli_type), 
  
  d %>% 
  filter(!is.na(Modality)) %>% 
  mutate(auditory_stimulus_type = case_when(
    Modality == "Speech" ~ "natural", 
    TRUE ~ "artificial"
  ))
)
```



### running models 

```{r eval=FALSE, include=FALSE}


best_fit_auditory_stimuli_df <- get_best_fit_model_with_mod(auditory_stimulus_d, "auditory_stimulus_type")
saveRDS(best_fit_auditory_stimuli_df, here("cached_data/best_fit_real_aud_df.Rds"))

```

### significant models 

```{r}
best_fit_real_aud_df %>% 
  filter(grepl("auditory_stimulus_typenatural", term)) %>% 
  filter(p.value < 0.05) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```
### visualization

```{r}
auditory_stimulus_d %>% 
   filter(mean_age_months < 36) %>% 
  ggplot(aes(x = mean_age_months, y = d_calc, color = auditory_stimulus_type)) + 
  geom_point(alpha = .5) + 
  geom_smooth(method = "lm") + 
  facet_wrap(~ds_clean)+ 
  theme_few()
```



# 4. Major author 

```{r}
all_nested_d <-lapply(distinct(d, ds_clean) %>% pull(), 
       function(x){
         get_major_author_effect(d, x, author_d)
       }) %>% 
  bind_rows() %>% 
  group_by(ds_clean, major_author) %>% 
  nest() %>% 
  filter(!major_author %in% c("…SweigWilson", "deVilliers", "Fernald","GrafEstes"
                              )) %>% 
  filter(!is.na(major_author) & ds_clean != "Prosocial Agents")

```

```{r eval=FALSE, include=FALSE}
major_author_model_ds <- lapply(seq(1, nrow(all_nested_d)), 
            function(x){
              print(x)
              get_author_model_comparison(all_nested_d, all_nested_d$ds_clean[[x]], 
                                          all_nested_d$major_author[[x]])
            }) %>% 
  bind_rows()

saveRDS(major_author_model_ds, here("cached_data/major_author_model_ds.Rds"))
```

```{r eval=FALSE, include=FALSE}
major_author_model_fits_ds <-  lapply(seq(1, nrow(all_nested_d)), 
            function(x){
              print(x)
              get_author_model_fits(all_nested_d, all_nested_d$ds_clean[[x]], 
                                          all_nested_d$major_author[[x]])
            }) %>% 
  bind_rows()

saveRDS(major_author_model_fits_ds, here("cached_data/major_author_fits_ds.Rds"))

```


## get the best fit major author models

```{r}
major_author_fit_df <- readRDS(here("cached_data/major_author_model_ds.Rds"))

major_author_model_fit_df <- readRDS(here("cached_data/major_author_fits_ds.Rds"))
```

## add in prosocial agents

```{r}
pa_d <- d %>% filter(ds_clean == "Prosocial agents") %>% 
  mutate(by_major_author = case_when(Hamlin_Lab == 1 ~ "yes", TRUE ~ "no"))


pa_model_d <- broom::tidy(rma.mv(
    d_calc ~ mean_age_months + by_major_author, 
      V = d_var_calc, 
      random = ~ 1 | short_cite/unique_row, 
      data = pa_d)) %>% 
  mutate(dataset = "Prosocial agents")


```


```{r eval=FALSE, include=FALSE}
min_AICc_major_author <- major_author_fit_df %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) 

best_fit_major_author_model <- major_author_model_fit_df %>% 
    left_join(min_AICc_major_author, 
              by =  c("model_spec", "dataset", "model_spec_clean", "major_author")) %>% 
    filter(!is.na(REML)) %>% 
    bind_rows(pa_model_d)



saveRDS(best_fit_major_author_model, here("cached_data/best_fit_major_author_df.Rds"))
```


## get the significant value 

```{r}
best_fit_major_author_df %>% 
  filter(grepl("by_major_author", term), 
         p.value < 0.05) 
```

```{r}
best_fit_major_author_df %>% 
  filter(grepl("by_major_author", term), 
         p.value < 0.05) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```

## visualization: not sure what's the best way to do it



# 5. Conflict between best fit models 

## AICc vs BIC

```{r}
aic_winner_df <- age_modesl_df %>% 
  filter(ic == "AICc") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) %>% 
  select(dataset, model_spec_clean, REML) %>% 
  rename(AICc_winner = model_spec_clean, 
         AIC_val = REML)

bic_winner_df <-  age_modesl_df %>% 
  filter(ic == "BIC") %>% 
  group_by(dataset) %>% 
  filter(REML == min(REML)) %>% 
  select(dataset, model_spec_clean, REML) %>% 
  rename(BIC_winner = model_spec_clean, 
         BIC_val = REML)

aic_winner_df %>% 
  left_join(bic_winner_df, by = c("dataset")) %>% 
  filter(!AICc_winner == BIC_winner) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```

## AICc: age vs age + moderators


some discrepancies but not a lot 
some have improvements and others do not 



```{r}
bind_rows(best_fit_bmeasure_model_df %>% mutate(add_mod = "bmeasure"), 
          best_fit_exposure_phase_df %>% mutate(add_mod = "exp_phase"), 
          best_fit_real_visual_df %>% mutate(add_mod = "visual_stimulus"),
          best_fit_real_aud_df %>% mutate(add_mod = "auditory_stimulus"), 
          best_fit_major_author_df %>% mutate(add_mod = "by_major_author")) %>% 
  select(dataset, model_spec_clean, REML, add_mod, major_author) %>% 
  distinct(dataset, model_spec_clean, REML, add_mod, major_author) %>% 
  rename(
    add_mod_winner_model = model_spec_clean, 
    add_mod_AICc = REML
  ) %>% 
  left_join(age_modesl_df %>% 
              filter(ic == "AICc") %>%
              group_by(dataset) %>% 
              filter(REML == min(REML)) %>% 
              rename(age_AICc = REML, 
                     age_winner_model = model_spec_clean)
              , by = c("dataset")) %>% 
  select(dataset, add_mod, add_mod_winner_model, add_mod_AICc, 
         age_AICc, age_winner_model) %>% 
  filter(add_mod_winner_model != age_winner_model) %>% 
  kableExtra::kable(digits = .2) %>%  
  kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```
